week-1-api-security/day-06-api-regression-harness/pyproject.toml
[tool.poetry]
name = "day-06-api-regression-harness"
version = "1.0.0"
description = "Data-driven API regression harness — YAML fixtures, parallel execution, HTML diff reporter, Allure integration"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python   = "^3.11"
httpx    = "^0.27"
pyyaml   = "^6.0"
fastapi  = "^0.110"
uvicorn  = {extras = ["standard"], version = "^0.27"}
pydantic = {extras = ["email"], version = "^2.6"}

[tool.poetry.dev-dependencies]
pytest     = "^8.1"
pytest-cov = "^5.0"

[tool.pytest.ini_options]
testpaths     = ["regression/tests"]
log_cli       = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["regression"]
omit   = ["*/tests/*"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
